Rename & move 'helpers.js' into 'helpers/everythingElse.js'
