+++ src/helpers/dataStructures.js
@@ -4,7 +4,7 @@ import {
   convertToISO8061,
   computeDayDifference,
   pureNumericalHourForm
-} from "/src/helpers.js";
+} from "/src/helpers/everythingElse.js";
 
 // FOR HANDLING A WEEK WORTH OF NEW TASKS
 // NOTE: by definition of these tasks being located on a different week, this will not
