+++ src/helpers/subtreeDragDrop.js
@@ -5,7 +5,7 @@ import {
   getDateInDDMMYYYY, 
   getTimeInHHMM,
   computeDayDifference
-} from '/src/helpers.js'
+} from '/src/helpers/everythingElse.js'
 import db from '/src/db.js'
 
 // break the parent relation only if necessary, so that the task can validly belong in the new todo list
