+++ src/lib/BedtimePopup.svelte
@@ -54,7 +54,7 @@
   import { createEventDispatcher, onMount, onDestroy, tick } from 'svelte'
   import _ from 'lodash'
 
-  import { getDateOfToday, getRandomID, clickOutside } from '/src/helpers.js'
+  import { getDateOfToday, getRandomID, clickOutside } from '/src/helpers/everythingElse.js'
   import { getFunctions, httpsCallable } from "firebase/functions"
   import { updateDoc, doc, getFirestore } from 'firebase/firestore'
   import { user } from '/src/store.js'
