+++ src/lib/ExperimentalCanvas.svelte
@@ -10,7 +10,7 @@
   import { onMount } from 'svelte'
   import BasicWhiteboard from '$lib/BasicWhiteboard.svelte'
   import { setFirestoreDoc } from '/src/crud.js'
-  import { getRandomID } from '/src/helpers.js'
+  import { getRandomID } from '/src/helpers/everythingElse.js'
   import { user } from '/src/store.js'
 
   function saveDoodleIconToDatabase (dataURL) {
