+++ src/lib/MobileMode.svelte
@@ -116,7 +116,7 @@
     getDateInDDMMYYYY,
     checkTaskObjSchema,
     sortByUnscheduledThenByOrderValue
-  } from '/src/helpers.js'
+  } from '/src/helpers/everythingElse.js'
   import { 
     user, 
     tasksScheduledOn,
