+++ src/lib/MyJSDatePicker.svelte
@@ -12,7 +12,7 @@
 <script>
   import 'js-datepicker/dist/datepicker.min.css'
   import { onMount, createEventDispatcher } from 'svelte'
-  import { getDateInMMDD, convertMMDDToDateClassObject } from '/src/helpers.js'
+  import { getDateInMMDD, convertMMDDToDateClassObject } from '/src/helpers/everythingElse.js'
 
   export let MMDD 
   export let YYYY
