+++ src/lib/ReusablePhotoTaskElement.svelte
@@ -97,7 +97,7 @@
 <script>
  // Assumes `task` is hydrated
  import { createEventDispatcher } from 'svelte'
- import { getTrueY } from '/src/helpers.js'
+ import { getTrueY } from '/src/helpers/everythingElse.js'
  import { yPosWithinBlock, whatIsBeingDragged, whatIsBeingDraggedID, whatIsBeingDraggedFullObj } from '/src/store.js'
  import ReusableCheckbox from '$lib/ReusableCheckbox.svelte'
  import { lazyCallable } from '/src/helpers/actions.js'
