--- src/storeSchema.js
@@ -1,6 +1,0 @@
-store = {
-
-    user: {
-        uid: ''
-    }
-}
